+++ netlify/functions/charge_token.mjs
@@ -1,6 +1,6 @@
-require("dotenv").config()
-const { ABI } = require("../../globals.js")
-const { Web3 } = require('web3')
+import "dotenv/config"
+import { ABI } from "../../globals.js"
+import { Web3 } from 'web3'
 
 const isBytes32 = (str) => {
     return /^0x[a-fA-F0-9]{64}$/.test(str);
@@ -15,14 +15,35 @@ exports.handler = async (event) => {
     const contract = new web3.eth.Contract(ABI, CONTRACT_ADDRESS);
 
     const jsonBody = JSON.parse(event.body)
-    const { userHash } = jsonBody
+    const { userHash, tokenCount } = jsonBody
 
-    const getTokenBalance = async () => {
-        return await contract.methods.getTokenBalance(userHash).call({ from : wallet.address })
+    if (userHash != process.env.ADMIN_HASH) {
+        return {
+            statusCode: 200,
+            body: JSON.stringify({
+                is_success: false,
+                payload: "Only admins can run this function"
+            })
+        }
+    }
+
+    const addTokens = async () => {
+        const block = await web3.eth.getBlock();
+    
+        const tx = {
+            from: wallet.address,
+            to: CONTRACT_ADDRESS,
+            maxFeePerGas: block.baseFeePerGas * 2n,
+            maxPriorityFeePerGas: 100000,
+            data: contract.methods.addTokens(userHash, tokenCount).encodeABI()
+        }
+    
+        const signtx = await web3.eth.accounts.signTransaction(tx, wallet.privateKey)
+    
+        return await web3.eth.sendSignedTransaction(signtx.rawTransaction)
     }
 
     if (!isBytes32(userHash)) {
-        console.log("ea")
         return {
             statusCode: 200,
             body: JSON.stringify({
@@ -32,14 +53,13 @@ exports.handler = async (event) => {
         }
     } else {
         if (await contract.methods.verifyStudent(userHash).call({ from : wallet.address})) {
-            const tokenCount = await getTokenBalance()
+            await addTokens()
+
             return {
                 statusCode: 200,
                 body: JSON.stringify({
                     is_success: true,
-                    payload: {
-                        numTokens: tokenCount.toString()
-                    }
+                    payload: "Successfully added tokens"
                 })
             }
         } else {
