+++ netlify/functions/register_candidates.mjs
@@ -1,13 +1,14 @@
-require("dotenv").config()
+import { db } from "../../globals.js"
+import { ref, onValue, set } from "firebase/database"
 
 exports.handler = async (event) => {
+    const messageRef = ref(db, "/register_candidates")
+
     const jsonBody = JSON.parse(event.body)
 
     const { adminHash } = jsonBody
 
     if (adminHash != process.env.ADMIN_HASH) {
-        console.log(process.env.ADMIN_HASH)
-        console.log(adminHash)
         return {
             statusCode: 200,
             body: JSON.stringify({
@@ -17,11 +18,21 @@ exports.handler = async (event) => {
         }
     }
 
+    const candidates = await new Promise((resolve, reject) => {
+        onValue(messageRef, snapshot => {
+            const candidates = snapshot.val()
+
+            resolve(candidates ?? {})
+        })
+    })
+
     return {
         statusCode: 200,
         body: JSON.stringify({
             is_success: true,
-            payload: "Success!"
+            payload: {
+                candidates: candidates ?? {}
+            }
         })
     }
 }
